Housekeeping: - Move Controller to own file - Clean up comments - Minor refactoring in View
